backend/requirements.txt
Flask-RESTful==0.3.10
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
firebase-admin==6.6.0
flask-redis